requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
bson==0.5.8
cffi==1.14.0
cryptography==2.9.2
dataclasses==0.6
Django==3.0.7
django-mock-queries==2.1.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
djongo==1.3.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.2
model-bakery==1.0.2
mongomock==3.19.0
more-itertools==8.2.0
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pylint==2.5.2
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
sentinels==1.0.0
six==1.14.0
sqlparse==0.2.4
toml==0.10.0
wcwidth==0.1.9
wrapt==1.12.1
